Added normal distribution to random number calculation

+++ pointset.h
@@ -22,8 +22,8 @@ public:
     // points at the nearest end are extrapolated.
     double interpolate(double x) const;
 
-    Point& operator [] (size_t idx) { return points_[idx]; }
-    const Point& operator [] (size_t idx) const { return points_[idx]; }
+    Point& operator [] (int idx) { return points_[idx]; }
+    const Point& operator [] (int idx) const { return points_[idx]; }
 
 protected:
     static double interpolate(double x1, double y1, double x2, double y2, double x);
